% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2pptx.R
\name{add_anyplot}
\alias{add_anyplot}
\title{Add a ggplot or a plot to the Microsoft Office Document}
\usage{
add_anyplot(
  doc,
  x = NULL,
  plottype = "auto",
  left = 1,
  top = 2,
  width = 8,
  height = 5.5
)
}
\arguments{
\item{doc}{A document object}

\item{x}{An object of class ggplot2 or a string encoding plot or ggplot}

\item{plottype}{character  One of c("auto","plot","ggplot","emf")}

\item{left}{left margin}

\item{top}{top margin}

\item{width}{desired width of the plot}

\item{height}{desired height of the plot}
}
\description{
Add a ggplot or a plot to the Microsoft Office Document
}
