% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf2pptx.R
\name{add_emf}
\alias{add_emf}
\title{Add a ggplot or a plot to the Microsoft Office Document}
\usage{
add_emf(doc, x, preprocessing = "", left = 1, top = 1, width = 8, height = 5.5)
}
\arguments{
\item{doc}{A document object}

\item{x}{A string encoding plot or ggplot}

\item{preprocessing}{A string of R code or ""}

\item{left}{left margin}

\item{top}{top margin}

\item{width}{desired width of the plot}

\item{height}{desired height of the plot}
}
\description{
Add a ggplot or a plot to the Microsoft Office Document
}
