% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ggplot.R
\name{add_ggplot}
\alias{add_ggplot}
\title{Add ggplot into a document object}
\usage{
add_ggplot(mydoc, code = "", top = 2)
}
\arguments{
\item{mydoc}{A document object}

\item{code}{R code for table}

\item{top}{top position of plot}
}
\value{
a document object
}
\description{
Add ggplot into a document object
}
\examples{
require(rrtable)
require(ggplot2)
require(officer)
require(magrittr)
code <- "ggplot(mtcars, aes(x = mpg , y = wt)) + geom_point()"
read_pptx() \%>\% add_text(title="ggplot") \%>\% add_ggplot(code=code)
}
