% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2ggplots.R
\name{add_self}
\alias{add_self}
\title{add self data to document}
\usage{
add_self(mydoc, data)
}
\arguments{
\item{mydoc}{A document object}

\item{data}{a data.frame}
}
\description{
add self data to document
}
