% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2ggplots.R
\name{add_text}
\alias{add_text}
\title{Add text to document}
\usage{
add_text(
  mydoc,
  title = "",
  text = "",
  code = "",
  preprocessing = "",
  echo = FALSE,
  eval = FALSE,
  style = "Normal",
  landscape = FALSE
)
}
\arguments{
\item{mydoc}{A document object}

\item{title}{An character string as a plot title}

\item{text}{text string to be added}

\item{code}{An R code string}

\item{preprocessing}{preprocessing}

\item{echo}{logical Whether or not show R code}

\item{eval}{logical whether or not evaluate the R code}

\item{style}{text style}

\item{landscape}{Logical. Whether or not make a landscape section.}
}
\description{
Add text to document
}
