% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2ggplots.R
\name{add_text2hyperlink}
\alias{add_text2hyperlink}
\title{Add hyperlink text}
\usage{
add_text2hyperlink(mydoc, text)
}
\arguments{
\item{mydoc}{A document object}

\item{text}{text string to be added}
}
\description{
Add hyperlink text
}
