% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2pdf.R
\name{adjustWidth}
\alias{adjustWidth}
\title{adjust width of data.frame}
\usage{
adjustWidth(data, width = 80, min = 10, aim = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{width}{total desired wdth}

\item{min}{minimum width of each column}

\item{aim}{desired column lengths}
}
\description{
adjust width of data.frame
}
