% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFlextable.R
\name{as.mynumeric}
\alias{as.mynumeric}
\title{coerce an object of type "numeric"}
\usage{
as.mynumeric(x)
}
\arguments{
\item{x}{A vector}
}
\description{
coerce an object of type "numeric"
}
\examples{
x=c("1,200","2","3.5")
x=factor(3:1)
x=c(1:3,"tt")
as.mynumeric(x)
}
