% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncProgressBar.R
\name{asyncProgressBar}
\alias{asyncProgressBar}
\title{Asynchronous progress bar sever function}
\usage{
asyncProgressBar(second = 2, message = NULL, interval = 100)
}
\arguments{
\item{second}{Numeric The timer of progress bar in seconds}

\item{message}{character Message of progress bar}

\item{interval}{numeric milliseconds}
}
\description{
Asynchronous progress bar sever function
}
