% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code2pptx.R
\name{code2docx}
\alias{code2docx}
\title{Save plot/ggplot code to Microsoft Powerpoint format}
\usage{
code2docx(...)
}
\arguments{
\item{...}{further arguments to be passed to code2office}
}
\description{
Save plot/ggplot code to Microsoft Powerpoint format
}
\examples{
\dontrun{
code2docx(plot(iris))
require(ggplot2)
gg=ggplot(data=mtcars,aes(x=wt,y=mpg))+geom_point()
code2docx(ggobj=gg)
}
}
