% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2pdf.R
\name{data2docx2}
\alias{data2docx2}
\title{Make a word file with a data.frame}
\usage{
data2docx2(...)
}
\arguments{
\item{...}{further arguments to be passed to data2HTML}
}
\description{
Make a word file with a data.frame
}
