% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2HTML.R
\name{data2HTML}
\alias{data2HTML}
\title{Make a HTML5 file with a data.frame}
\usage{
data2HTML(
  data,
  preprocessing = "",
  path = ".",
  filename = "report.HTML",
  rawDataName = NULL,
  rawDataFile = "rawData.RDS",
  type = "HTML",
  vanilla = FALSE,
  echo = TRUE,
  showself = FALSE
)
}
\arguments{
\item{data}{A data.frame}

\item{preprocessing}{A character string of R code}

\item{path}{A name of destination file path}

\item{filename}{A name of destination file}

\item{rawDataName}{The name of the rawData}

\item{rawDataFile}{The name of the rawData file which the data are to be read from.}

\item{type}{character "HTML" or "pdf"}

\item{vanilla}{logical. Whether or not make vanilla table}

\item{echo}{Logical. Whether or not show R code of plot and table}

\item{showself}{Logical. Whether or not show R code for the paragraph}
}
\description{
Make a HTML5 file with a data.frame
}
\examples{
\donttest{
library(moonBook)
library(rrtable)
library(ggplot2)
data2HTML(sampleData2)
}
}
