% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2office.R
\name{data2office}
\alias{data2office}
\title{convert data to pptx file}
\usage{
data2office(
  data,
  preprocessing = "",
  path = ".",
  filename = "Report",
  format = "pptx",
  width = 7,
  height = 5,
  units = "in",
  res = 300,
  rawDataName = NULL,
  rawDataFile = "rawData.RDS",
  vanilla = FALSE,
  echo = FALSE,
  landscape = FALSE,
  showself = FALSE,
  out = NULL
)
}
\arguments{
\item{data}{A document object}

\item{preprocessing}{A string}

\item{path}{A name of destination file path}

\item{filename}{File name}

\item{format}{desired format. choices are "pptx" or "docx"}

\item{width}{the width of the device.}

\item{height}{the height of the device.}

\item{units}{The units in which height and width are given. Can be px (pixels, the default), in (inches), cm or mm.}

\item{res}{The nominal resolution in ppi which will be recorded in the bitmap file, if a positive integer. Also used for units other than the default, and to convert points to pixels.}

\item{rawDataName}{raw Data Name}

\item{rawDataFile}{raw Data File}

\item{vanilla}{logical. Whether or not make vanilla table}

\item{echo}{logical Whether or not show R code}

\item{landscape}{Logical. Whether or not make a landscape section.}

\item{showself}{Logical. Whether or not show R code for the paragraph}

\item{out}{An object or NULL}
}
\description{
convert data to pptx file
}
