% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2office.R
\name{data2pptx}
\alias{data2pptx}
\title{convert data to pptx file}
\usage{
data2pptx(...)
}
\arguments{
\item{...}{arguments to be passed to data2office()}
}
\description{
convert data to pptx file
}
\examples{
\dontrun{
library(rrtable)
library(moonBook)
library(ggplot2)
data2pptx(sampleData2,echo=TRUE)
}
}
