% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2pdf.R
\name{data2pptx2}
\alias{data2pptx2}
\title{Make a powerpoint file with a data.frame}
\usage{
data2pptx2(...)
}
\arguments{
\item{...}{further arguments to be passed to data2HTML}
}
\description{
Make a powerpoint file with a data.frame
}
\examples{
library(moonBook)
library(ztable)
library(ggplot2)
\donttest{
data2pptx2(sampleData2,echo=FALSE)
}
}
