% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2flextable.R
\name{df2flextable2}
\alias{df2flextable2}
\title{Make flextable with limited width}
\usage{
df2flextable2(df, mincol = 0.7, maxcol = 4, ...)
}
\arguments{
\item{df}{a data.frame}

\item{mincol}{minimum column width in inch}

\item{maxcol}{maximum column width in inch}

\item{...}{further arguments to be passed to df2flextable()}
}
\description{
Make flextable with limited width
}
