% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf2pptx.R
\name{emf2pptx}
\alias{emf2pptx}
\title{Save plot/ggplot to Microsoft Powerpoint format}
\usage{
emf2pptx(
  x,
  target = "Report",
  append = FALSE,
  title = "",
  type = "pptx",
  preprocessing = "",
  left = 1,
  top = 1,
  width = 8,
  height = 5.5
)
}
\arguments{
\item{x}{A string vector encoding plot or ggplot}

\item{target}{name of output file}

\item{append}{logical value}

\item{title}{Optional character vector of plot title}

\item{type}{"pptx" or "docx"}

\item{preprocessing}{A string of R code or ""}

\item{left}{left margin}

\item{top}{top margin}

\item{width}{desired width of the plot}

\item{height}{desired height of the plot}
}
\description{
Save plot/ggplot to Microsoft Powerpoint format
}
\examples{
require(ggplot2)
emf2pptx("ggplot(data=iris,aes(x=Sepal.Length))+geom_density()")
}
