% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2pdf.R
\name{file2plotzip}
\alias{file2plotzip}
\title{read data file and make a zip file with plots}
\usage{
file2plotzip(file, selected = NULL, ...)
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{selected}{A numeric vector or NULL(default). If specified, only selected data are printed.}

\item{...}{Further argument to be passed to data2plotzip()}
}
\description{
read data file and make a zip file with plots
}
