% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2pdf.R
\name{file2pptx}
\alias{file2pptx}
\title{read data file and make a pptx file}
\usage{
file2pptx(file, selected = NULL, ...)
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{selected}{A numeric vector or NULL(default). If specified, only selected data are printed.}

\item{...}{Further argument to be passed to data2pptx()}
}
\description{
read data file and make a pptx file
}
