% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image2pptx.R
\name{image2docx}
\alias{image2docx}
\title{Save plot/ggplot to Microsoft Word format}
\usage{
image2docx(...)
}
\arguments{
\item{...}{further arguments to be passed to image2office}
}
\description{
Save plot/ggplot to Microsoft Word format
}
\examples{
\donttest{
require(ggplot2)
x<-ggplot(iris,aes(x=Sepal.Length))+geom_histogram()
image2docx(x)
image2docx(x="plot(iris)",title="A ggplot",append=TRUE)
p2="ggplot(iris,aes(x=Sepal.Length,y=Sepal.Width))+geom_point()"
image2docx(p2,append=TRUE)
}
}
