% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image2pptx.R
\name{image2pptx}
\alias{image2pptx}
\title{Save plot/ggplot to Microsoft Powerpoint format}
\usage{
image2pptx(...)
}
\arguments{
\item{...}{further arguments to be passed to image2office}
}
\description{
Save plot/ggplot to Microsoft Powerpoint format
}
\examples{
\dontrun{
require(ggplot2)
x<-ggplot(iris,aes(x=Sepal.Length))+geom_histogram()
image2pptx(x)
x="plot(iris)"
image2pptx(x,title="A plot",append=TRUE)
p2="ggplot(iris,aes(x=Sepal.Length,y=Sepal.Width))+geom_point()"
image2pptx(p2,append=TRUE)
}
}
