% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2pptx.R
\name{is_ggplot}
\alias{is_ggplot}
\title{Reports whether plotstring encode a ggplot object}
\usage{
is_ggplot(plotstring)
}
\arguments{
\item{plotstring}{A character}
}
\description{
Reports whether plotstring encode a ggplot object
}
\examples{
require(ggplot2)
is_ggplot("plot(iris)")
is_ggplot("ggplot(iris,aes(x=Sepal.Length))+geom_histogram()")
}
