% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2plotzip.R
\name{is_ggsurvplot}
\alias{is_ggsurvplot}
\title{Reports whether plotstring encode a ggsurvplot object}
\usage{
is_ggsurvplot(x, preprocessing = "", out = NULL)
}
\arguments{
\item{x}{A character encoding a plot}

\item{preprocessing}{preprocessing}

\item{out}{An object or NULL}
}
\description{
Reports whether plotstring encode a ggsurvplot object
}
