% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLineTable.R
\name{lfData}
\alias{lfData}
\title{Seperate linefeed as a separate row in a data.frame}
\usage{
lfData(data)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Seperate linefeed as a separate row in a data.frame
}
