% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLineTable.R
\name{multiLineCount}
\alias{multiLineCount}
\title{Count multiline in a data.frame}
\usage{
multiLineCount(data, colno)
}
\arguments{
\item{data}{a data.frame}

\item{colno}{column number to separate}
}
\description{
Count multiline in a data.frame
}
