% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLineTable.R
\name{multiLineData}
\alias{multiLineData}
\title{make multiline data}
\usage{
multiLineData(data, colno = 4)
}
\arguments{
\item{data}{a data.frame}

\item{colno}{column number to separate}
}
\description{
make multiline data
}
