% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLineTable.R
\name{multiLineTable}
\alias{multiLineTable}
\title{Make flextable with multiline}
\usage{
multiLineTable(data, colno = 4)
}
\arguments{
\item{data}{a data.frame}

\item{colno}{column number to separate}
}
\value{
a flextable
}
\description{
Make flextable with multiline
}
\examples{
multiLineTable(sampleData3,4)
}
