% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFlextable.R
\name{myFlextable}
\alias{myFlextable}
\title{Make flextable with a data.frame}
\usage{
myFlextable(df, numericCol = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{numericCol}{Numeric. Columns to be treated as numeric}
}
\description{
Make flextable with a data.frame
}
