% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2office.R
\name{mygrep}
\alias{mygrep}
\title{grep string in all files in subdirectory}
\usage{
mygrep(x, file = "*")
}
\arguments{
\item{x}{string}

\item{file}{files to seek}
}
\description{
grep string in all files in subdirectory
}
