% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2pptx.R
\name{open_doc}
\alias{open_doc}
\title{Make/open office document with file name}
\usage{
open_doc(target = "Report", type = "pptx", append = FALSE)
}
\arguments{
\item{target}{name of output file}

\item{type}{"pptx" or "docx"}

\item{append}{logical}
}
\description{
Make/open office document with file name
}
