% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2flextable.R
\name{p2character}
\alias{p2character}
\title{Change p value to string}
\usage{
p2character(x, digits = 3)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Change p value to string
}
\examples{
x=c(0.000001,NA,0.1234567,0.00123,0.000123)
p2character(x)
p2character(x,digits=4)
}
