% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\name{pickerInput3}
\alias{pickerInput3}
\title{Side by side pickerInput}
\usage{
pickerInput3(...)
}
\arguments{
\item{...}{Further arguments to be passed to pickerInput}
}
\description{
Side by side pickerInput
}
