% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2pptx.R
\name{plot2office}
\alias{plot2office}
\title{Save plot/ggplot to Microsoft Powerpoint format}
\usage{
plot2office(
  p = NULL,
  plotstring = NULL,
  target = "plot",
  append = FALSE,
  title = "",
  type = "pptx",
  left = 1,
  top = 1,
  width = NULL,
  height = NULL,
  aspectr = NULL
)
}
\arguments{
\item{p}{An object of class ggplot2 or NULL}

\item{plotstring}{A character}

\item{target}{name of output file}

\item{append}{logical value}

\item{title}{Optional character of plot title}

\item{type}{"pptx" or "docx"}

\item{left}{left margin}

\item{top}{top margin}

\item{width}{desired width of the plot}

\item{height}{desired height of the plot}

\item{aspectr}{desired aspect ratio of the plot}
}
\description{
Save plot/ggplot to Microsoft Powerpoint format
}
