% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Rcode.R
\name{Rcode2df}
\alias{Rcode2df}
\title{Make a data.frame with character strings encoding R code}
\usage{
Rcode2df(result, preprocessing, eval = TRUE, out = NULL)
}
\arguments{
\item{result}{character strings encoding R code}

\item{preprocessing}{character strings encoding R code as a preprocessing}

\item{eval}{logical. Whether or not evaluate the code}

\item{out}{An object or NULL}
}
\description{
Make a data.frame with character strings encoding R code
}
