% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Rcode.R
\name{Rcode2df2}
\alias{Rcode2df2}
\title{Make a data.frame with character strings encoding R code}
\usage{
Rcode2df2(result, eval = TRUE)
}
\arguments{
\item{result}{character strings encoding R code}

\item{eval}{logical. Whether or not evaluate the code}
}
\description{
Make a data.frame with character strings encoding R code
}
