% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Rcode.R
\name{Rcode2docx}
\alias{Rcode2docx}
\title{Save R code to Microsoft Word format}
\usage{
Rcode2docx(...)
}
\arguments{
\item{...}{further arguments to be passed to plot2office}
}
\description{
Save R code to Microsoft Word format
}
\examples{
\dontrun{
code="summary(lm(mpg~hp+wt,data=mtcars))"
Rcode2docx(code=code,title="R code to Word")
}
}
