% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Rcode.R
\name{Rcode2office}
\alias{Rcode2office}
\title{Make R code slide}
\usage{
Rcode2office(
  code,
  preprocessing = "",
  title = "",
  type = "pptx",
  target = "Report",
  append = FALSE
)
}
\arguments{
\item{code}{A character string encoding R codes}

\item{preprocessing}{A character string of R code as a preprocessing}

\item{title}{A character}

\item{type}{desired format. choices are "pptx" or "docx"}

\item{target}{name of output file}

\item{append}{logical}
}
\description{
Make R code slide
}
\examples{
\donttest{
code="summary(lm(mpg~hp+wt,data=mtcars))"
Rcode2office(code=code)
}
}
