% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Rcode.R
\name{Rcode2pptx}
\alias{Rcode2pptx}
\title{Save R code to Microsoft Powerpoint format}
\usage{
Rcode2pptx(...)
}
\arguments{
\item{...}{further arguments to be passed to plot2office}
}
\description{
Save R code to Microsoft Powerpoint format
}
\examples{
\donttest{
code="summary(lm(mpg~hp+wt,data=mtcars))"
Rcode2pptx(code=code,title="R code to pptx")
}
}
