% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSVComment.R
\name{readCSVComment}
\alias{readCSVComment}
\title{Read a csv file with comment}
\usage{
readCSVComment(file)
}
\arguments{
\item{file}{A path for destination file}
}
\description{
Read a csv file with comment
}
