% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxList.R
\docType{data}
\name{sampleData3}
\alias{sampleData3}
\title{Sample data for pptxList
A dataset containing five objects for reproducible research}
\format{A data frame with 5 rows and three columns
\describe{
   \item{type}{type of data}
   \item{title}{title of data}
   \item{text}{text}
   \item{code}{R code of data}
   \item{option}{option for R code}
}}
\usage{
sampleData3
}
\description{
Sample data for pptxList
A dataset containing five objects for reproducible research
}
\keyword{datasets}
