% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2office.R
\name{table2docx}
\alias{table2docx}
\title{Export data.frame or statistical output to Microsoft Word format}
\usage{
table2docx(...)
}
\arguments{
\item{...}{further arguments to be passed to table2office}
}
\description{
Export data.frame or statistical output to Microsoft Word format
}
\examples{
\donttest{
require(moonBook)
x=mytable(Dx~.,data=acs)
table2docx(x)
table2docx(head(iris),title="head(iris)",append=TRUE,vanilla=FALSE)
fit=lm(mpg~wt*hp,data=mtcars)
table2docx(fit,title="Linear regression",append=TRUE,vanilla=TRUE)
fit2=aov(yield ~ block + N * P + K, data = npk)
table2docx(fit2,title="Linear regression",append=TRUE,vanilla=TRUE)
}
}
