% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2office.R
\name{table2office}
\alias{table2office}
\title{Export data.frame or statistical output to a table in Microsoft Office}
\usage{
table2office(
  x = NULL,
  target = "Report",
  append = FALSE,
  title = "",
  vanilla = FALSE,
  echo = FALSE,
  preprocessing = "",
  type = "pptx",
  landscape = FALSE,
  left = 1,
  top = 1
)
}
\arguments{
\item{x}{An object or string}

\item{target}{name of output file}

\item{append}{logical value}

\item{title}{Optional character of plot title}

\item{vanilla}{A logical}

\item{echo}{logical}

\item{preprocessing}{A character string}

\item{type}{"pptx" or "docx"}

\item{landscape}{logical}

\item{left}{left margin}

\item{top}{top margin}
}
\description{
Export data.frame or statistical output to a table in Microsoft Office
}
