% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2office.R
\name{table2pptx}
\alias{table2pptx}
\title{Export data.frame or statistical output to Microsoft Powerpoint format}
\usage{
table2pptx(...)
}
\arguments{
\item{...}{further arguments to be passed to table2office}
}
\description{
Export data.frame or statistical output to Microsoft Powerpoint format
}
\examples{
require(moonBook)
x=mytable(Dx~.,data=acs)
table2pptx(x)
table2pptx(head(iris),title="head(iris)",append=TRUE,vanilla=FALSE)
fit=lm(mpg~wt*hp,data=mtcars)
table2pptx(fit,title="Linear regression",append=TRUE,vanilla=TRUE)
fit2=aov(yield ~ block + N * P + K, data = npk)
table2pptx(fit2,title="Linear regression",append=TRUE,vanilla=TRUE)
}
