% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2pdf.R
\name{tensiSplit2}
\alias{tensiSplit2}
\title{Split string vectors with desired length with exdent}
\usage{
tensiSplit2(x, ...)
}
\arguments{
\item{x}{A string vector}

\item{...}{further argument to be passed to tensiSplit}
}
\description{
Split string vectors with desired length with exdent
}
