% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncProgressBar.R
\name{timeProgress}
\alias{timeProgress}
\title{Server function of asyncProgressBar Module}
\usage{
timeProgress(
  input,
  output,
  session,
  second = 2,
  message = NULL,
  interval = 100
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{second}{Numeric The timer of progress bar in seconds}

\item{message}{character Message of progress bar}

\item{interval}{numeric milliseconds}
}
\description{
Server function of asyncProgressBar Module
}
