% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2pdf.R
\name{ztable2}
\alias{ztable2}
\title{make latex table with ztable}
\usage{
ztable2(data, aim = NULL, type = "latex", ...)
}
\arguments{
\item{data}{a data.frame}

\item{aim}{desired column lengths}

\item{type}{output type}

\item{...}{further argument to be passed to ztable}
}
\description{
make latex table with ztable
}
\examples{
require(magrittr)
require(ztable)
ztable2(sampleData3)
}
