% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2flextable.R
\name{ztable2flextable}
\alias{ztable2flextable}
\title{Convert ztable to flextable}
\usage{
ztable2flextable(z, ...)
}
\arguments{
\item{z}{An object of class ztable}

\item{...}{Further argument to be passed to df2flextable}
}
\value{
an object of class flextable
}
\description{
Convert ztable to flextable
}
