% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLineTable.R
\name{ztable3}
\alias{ztable3}
\title{Make ztable with desired width}
\usage{
ztable3(df, cwidth = NULL, width = 80)
}
\arguments{
\item{df}{a data.frame}

\item{cwidth}{desired column width}

\item{width}{desired table width in column}
}
\description{
Make ztable with desired width
}
