% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Rcode.R
\name{add_Rcode}
\alias{add_Rcode}
\title{Make a R code slide into a document object}
\usage{
add_Rcode(mydoc, code, preprocessing = "", format = "pptx")
}
\arguments{
\item{mydoc}{A document object}

\item{code}{A character string encoding R codes}

\item{preprocessing}{A character string of R code as a preprocessing}

\item{format}{desired format. choices are "pptx" or "docx"}
}
\value{
a document object
}
\description{
Make a R code slide into a document object
}
\examples{
library(rrtable)
library(magrittr)
library(officer)
code="summary(lm(mpg~hp+wt,data=mtcars))"
read_pptx() \%>\% add_text(title="Regression Analysis") \%>\%
   add_Rcode(code) \%>\% print(target=paste0(tempdir(),"/","test.pptx"))
}
