% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2pdf.R
\name{data2pdf}
\alias{data2pdf}
\title{Make a pdf file with a data.frame}
\usage{
data2pdf(...)
}
\arguments{
\item{...}{further arguments to be passed to data2HTML}
}
\description{
Make a pdf file with a data.frame
}
\examples{
library(moonBook)
library(ztable)
library(ggplot2)
\donttest{
data2pdf(sampleData2)
}
}
