% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2pdf.R
\name{data2pptx2}
\alias{data2pptx2}
\title{Make a Powerpoint file with a data.frame}
\usage{
data2pptx2(...)
}
\arguments{
\item{...}{further arguments to be passed to data2HTML}
}
\description{
Make a Powerpoint file with a data.frame
}
