% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable2ztable.R
\name{flextable2ztable}
\alias{flextable2ztable}
\title{Convert flextable to ztable}
\usage{
flextable2ztable(ft, type = "html", ...)
}
\arguments{
\item{ft}{An object of class flextable}

\item{type}{"html" or "latex"}

\item{...}{Further argument to be passed to ztable}
}
\value{
an object of class ztable
}
\description{
Convert flextable to ztable
}
