% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2pptx.R
\name{plot2docx}
\alias{plot2docx}
\title{Save plot/ggplot to Microsoft Word format}
\usage{
plot2docx(...)
}
\arguments{
\item{...}{further arguments to be passed to plot2office}
}
\description{
Save plot/ggplot to Microsoft Word format
}
\examples{
\dontrun{
require(ggplot2)
x<-ggplot(iris,aes(x=Sepal.Length))+geom_histogram()
plot2docx(x)
plot2docx(x,title="A ggplot",append=TRUE)
p2=ggplot(iris,aes(x=Sepal.Length,y=Sepal.Width))+geom_point()
plot2docx(p2,append=TRUE)
plot2docx(x="plot(iris)",append=TRUE,title="plot(iris)")
plot2docx(x="ggplot(iris,aes(x=Sepal.Length))+geom_histogram()",append=TRUE)
}
}
