% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2pptx.R
\name{plot2pptx}
\alias{plot2pptx}
\title{Save plot/ggplot to Microsoft Powerpoint format}
\usage{
plot2pptx(...)
}
\arguments{
\item{...}{further arguments to be passed to plot2office}
}
\description{
Save plot/ggplot to Microsoft Powerpoint format
}
\examples{
\dontrun{
require(ggplot2)
x<-ggplot(iris,aes(x=Sepal.Length))+geom_histogram()
plot2pptx(x)
plot2pptx(x,title="A ggplot",append=TRUE)
p2=ggplot(iris,aes(x=Sepal.Length,y=Sepal.Width))+geom_point()
plot2pptx(p2,append=TRUE)
plot2pptx(x=c("plot(iris)","ggplot(iris,aes(x=Sepal.Length))+geom_histogram()"),
    append=TRUE,title=c("plot","ggplot"),echo=TRUE)
}
}
