% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2office.R
\name{table2pptx}
\alias{table2pptx}
\title{Export data.frame or statistical output to Microsoft Powerpoint format}
\usage{
table2pptx(...)
}
\arguments{
\item{...}{further arguments to be passed to table2office}
}
\description{
Export data.frame or statistical output to Microsoft Powerpoint format
}
\examples{
\donttest{
require(moonBook)
x="mytable(Dx~.,data=acs)"
table2pptx(x,title="mytable object",echo=TRUE)
table2pptx("head(iris)",title="data.Frame",append=TRUE,vanilla=FALSE,echo=TRUE)
x="fit<-lm(mpg~wt*hp,data=mtcars);fit"
table2pptx(x,title="Linear regression",append=TRUE,vanilla=TRUE,echo=TRUE)
fit2="aov(yield ~ block + N * P + K, data = npk)"
table2pptx(fit2,title="ANOVA",append=TRUE,vanilla=TRUE,echo=TRUE)
}
}
