% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{ztable2}
\alias{ztable2}
\title{Make ztable with desired width}
\usage{
ztable2(df, cwidth = NULL, width = 80, ...)
}
\arguments{
\item{df}{a data.frame}

\item{cwidth}{desired column width}

\item{width}{desired table width in column}

\item{...}{further argument to be passed to ztable()}
}
\description{
Make ztable with desired width
}
